+++ docs/make.jl
@@ -4,6 +4,9 @@ using ComplexityMeasures
 # Convert tutorial file to markdown
 import Literate
 Literate.markdown("src/tutorial.jl", "src"; credit = false)
+Literate.markdown("src/measure_count.jl", "src"; credit = false)
+
+import Documenter
 
 pages = [
     "index.md",
@@ -16,6 +19,7 @@ pages = [
     "examples.md",
     "devdocs.md",
     "references.md",
+    Documenter.hide("measure_count.md"),
 ]
 
 # For easier debugging when downloading from a specific branch.